comentários e ajustes

+++ servidor.py
@@ -13,6 +13,9 @@ clientes = []
 caminho_projeto = os.path.dirname(os.path.abspath("__file__"))
 
 
+"""
+Função main que inicializa o cliente, conectando ao cliente, e a thread
+"""
 def main():
     servidor = socket.socket(socket.AF_INET, socket.SOCK_STREAM) 
 
@@ -36,7 +39,15 @@ def main():
         thread = Thread(target=processar_msgs, args=[cliente])
         thread.start()
 
-#processa as msgs enviadas pelo cliente
+
+
+"""
+Função para a Thread, recebe as mensagens que o cliente envia
+    
+    :param parametro1: cliente.
+    :tipo parametro1: socket.
+
+"""
 def processar_msgs(cliente):
     while True:
         try:
@@ -45,11 +56,6 @@ def processar_msgs(cliente):
             msg = objeto["comando"]
             msg_nome_arquivo = objeto["dados"]
             
-            a, b = cliente.getsockname()
-            print (a)
-            print (b)
-            print(clientes)
-
             if msg == "listagem_servidor":
                 arquivos = listar_arquivos("SERVIDOR")
                 enviar_estrutura_msg("listagem_servidor", arquivos, cliente)
@@ -66,8 +72,11 @@ def processar_msgs(cliente):
                 time.sleep(2)
             
             elif msg == 'solicitar_download':
-                copiar_arquivo("SERVIDOR","CLIENTE",msg_nome_arquivo)
-                enviar_estrutura_msg("retorno", "Download realizado com êxito!", cliente)
+                msg_retorno = copiar_arquivo("SERVIDOR","CLIENTE",msg_nome_arquivo)
+                if msg_retorno == True:
+                    enviar_estrutura_msg("retorno", "Download realizado com êxito!", cliente)
+                else:
+                    enviar_estrutura_msg("retorno", msg_retorno, cliente)
                 time.sleep(2)
 
             elif msg == 'delecao':
@@ -86,8 +95,11 @@ def processar_msgs(cliente):
                 time.sleep(2)
             
             elif msg == 'solicitar_upload':
-                copiar_arquivo("CLIENTE","SERVIDOR",msg_nome_arquivo)
-                enviar_estrutura_msg("retorno", "Upload do arquivo realizado com êxito!", cliente)
+                msg_retorno = copiar_arquivo("CLIENTE","SERVIDOR",msg_nome_arquivo)
+                if msg_retorno == True:
+                    enviar_estrutura_msg("retorno", "Upload do arquivo realizado com êxito!", cliente)
+                else:
+                    enviar_estrutura_msg("retorno", msg_retorno, cliente)
                 time.sleep(2)
                           
 
@@ -97,19 +109,43 @@ def processar_msgs(cliente):
             main()
        
 
-#deletar clientes que não estão mais conectados
+"""
+Deleta os clientes inativos
+    
+    :param parametro1: cliente.
+    :tipo parametro1: socket.
+
+"""
 def delete_cliente(cliente):
     clientes.remove(cliente)
 
 
-#enviar msg para cliente
+"""
+Envia a mensagem para o cliente
+    
+    :param parametro1: cliente
+    :tipo parametro1: socket
+    :param parametro1: msg
+    :tipo parametro1: string
+
+"""
 def enviar_msgs(cliente, msg):
     try:
         cliente.send(msg.encode('utf-8'))
     except:
         delete_cliente(cliente)
 
+
 #funções de manipulação de arquivos --------------------------------
+        
+"""
+Procura os arquivos e organiza em uma string para serem listados
+
+    :param parametro1: pasta
+    :tipo parametro1: string.
+    :return: Mensagem com o nome dos arquivos.
+    :rtipo: string.
+"""
 def listar_arquivos(pasta):
     lista = []
     caminho_pasta = os.path.join(caminho_projeto,pasta)
@@ -121,12 +157,37 @@ def listar_arquivos(pasta):
     msg = tratar_vetor(lista)
     return msg
 
+
+"""
+Procura os arquivos e organiza em uma string para serem listados
+
+    :param parametro1: origem
+    :tipo parametro1: string.
+    :param parametro2: destino
+    :tipo parametro2: string.
+    :param parametro3: nome_arquivo
+    :tipo parametro3: string.
+    :return: Mensagem  de erro ou "True" para indicar sucesso.
+    :rtipo: string e booleano.
+"""
 def copiar_arquivo(origem, destino, nome_arquivo):
     caminho_origem = os.path.join(caminho_projeto, origem, nome_arquivo)
     caminho_destino = os.path.join(caminho_projeto, destino)
-    shutil.copy2(caminho_origem, caminho_destino)
+    if os.path.exists(caminho_origem):
+        shutil.copy2(caminho_origem, caminho_destino)
+        return True
+    else:
+        return "Não foi possível realizar a ação, verifique o nome do arquivo."
+    
 
+"""
+Procura os arquivos e organiza em uma string para serem listados
 
+    :param parametro1: nome_arquivo
+    :tipo parametro1: string.
+    :return: Mensagem .
+    :rtipo: string.
+"""
 def deletar_arquivo(nome_arquivo):
     caminho_origem = os.path.join(caminho_projeto, 'SERVIDOR')
     caminho_excluir = os.path.join(caminho_projeto, caminho_origem, nome_arquivo)
@@ -136,25 +197,54 @@ def deletar_arquivo(nome_arquivo):
     else:
         return "O arquivo não existe. Digite o nome do arquivo corretamente."
 
-def escrever_log(mensagem):
-    data_public = datetime.now()
-    data = data_public.strftime("%d/%m/%Y %H:%M")
-    arquivo = open ("logs.txt",'a')
-    arquivo.write(data + " : " + mensagem+"\n")
-    arquivo.close()
 
+"""
+Trata o vetor de arquivos recebido, transformando em uma string
+    
+    :param parametro1: msg.
+    :tipo parametro1: string.
+    :return: mensagem tratada .
+    :rtipo: string.
+
+"""
 def tratar_vetor(vetor):
     tratado = ",".join(vetor)
     return tratado 
 
+
+"""
+Função que deserializa uma mensagem depois de ser recebida via socket
+    :param parametro1: objeto.
+    :tipo parametro1: string.
+    :return: Objeto deserializado.
+    :rtipo: dicionario.
+"""
 def deserializar(objeto_recebido):
     objeto = pickle.loads(objeto_recebido)
     return objeto
 
+
+"""
+Função que serializa uma mensagem que não seja uma string simples para ser enviada via socket
+    :param parametro1: objeto.
+    :tipo parametro1: dicionário.
+    :return: Objeto serializado.
+    :rtipo: obj.
+"""
 def serializar (objeto):
     objeto_serializado = pickle.dumps(objeto)
     return objeto_serializado
 
+
+"""
+Envia a mensagem desejada no formato de dicionário, mas para isso é necessário os dados de comando e as informações
+    :param parametro1: comando.
+    :tipo parametro1: string.
+    :param parametro2: dados.
+    :tipo parametro2: string.
+    :param parametro3: cliente.
+    :tipo parametro3: socket.
+"""
 def enviar_estrutura_msg(comando, dados, cliente):
     try:
         msg = {"comando":comando,"dados":dados}
